chore: rename, refactor directories files
