+++ src/main/java/com/wap/cano_be/member/dto/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package com.wap.cano_be.member.domain;
+package com.wap.cano_be.member.dto;
 
 public record MemberRequestDto(
         String email,
